update: considering the ratio's rather than only raw pixel values
As we're working with raw pixel values the pixel threshold is not
working as the distance is varying from the camera. so I am considering
the ratio of the face top(forehead) and chin (upper chin) and getting
all the pixel values as a ratio by didiving by this constant distance
which will change in proportion to the distance from the camera of THE
user.